+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 
 - ✨ New button to stop all processes at once in the dashboard.
 - ✨ Improved the MacOS help menu with links to the documentation and changelog.
+- 🐞 Correctly handle "live update" log instructions.
 
 ## 1.2.1 - 2025-10-20